requirements.dev.txt
fastapi-cli==0.0.14

